Fix type errors
