+++ registry/lib/components/feeds/filter/FilterTypeSwitch.vue
@@ -12,8 +12,9 @@
 <script lang="ts">
 import { getComponentSettings } from '@/core/settings'
 import { VIcon } from '@/ui'
+import { FeedsFilterOptions } from './options'
 
-const { options } = getComponentSettings('feedsFilter')
+const { options } = getComponentSettings<FeedsFilterOptions>('feedsFilter')
 export default Vue.extend({
   components: {
     VIcon,
@@ -50,12 +51,13 @@ export default Vue.extend({
       if (!updateSettings) {
         return
       }
+      const optionKey = this.optionKey as 'types' | 'specialTypes'
       if (disabled) {
-        options[this.optionKey].push(this.type.id)
+        options[optionKey].push(this.type.id)
       } else {
-        const index = options[this.optionKey].indexOf(this.type.id)
+        const index = options[optionKey].indexOf(this.type.id)
         if (index !== -1) {
-          options[this.optionKey].splice(index, 1)
+          options[optionKey].splice(index, 1)
         }
       }
     },
