+++ registry/lib/components/feeds/filter/index.ts
@@ -1,10 +1,7 @@
-import {
-  defineOptionsMetadata,
-  defineComponentMetadata,
-  OptionsOfMetadata,
-} from '@/components/define'
+import { defineComponentMetadata } from '@/components/define'
 import { feedsCardsManager } from '@/components/feeds/api'
 import { feedsFilterPlugin } from './plugin'
+import { options } from './options'
 
 const entry = async () => {
   const { select } = await import('@/core/spin-query')
@@ -25,32 +22,6 @@ const entry = async () => {
   const { mountVueComponent } = await import('@/core/utils')
   leftPanel.insertAdjacentElement('afterbegin', mountVueComponent(FeedsFilterCard).$el)
 }
-
-const options = defineOptionsMetadata({
-  types: {
-    defaultValue: [] as number[],
-    displayName: '过滤动态类型',
-    hidden: true,
-  },
-  patterns: {
-    defaultValue: [] as string[],
-    displayName: '过滤关键词',
-    hidden: true,
-  },
-  sideCards: {
-    /** FIXME: 虽然代码里按照`number[]`, 但其实存进去的是`string[]`? */
-    defaultValue: [] as number[],
-    displayName: '过滤侧边栏',
-    hidden: true,
-  },
-  specialTypes: {
-    defaultValue: [] as number[],
-    displayName: '过滤特殊动态类型',
-    hidden: true,
-  },
-})
-
-export type FeedsFilterOptions = OptionsOfMetadata<typeof options>
 
 export const component = defineComponentMetadata({
   name: 'feedsFilter',
